reorganize all discretisation-related files to be under one folder
