+++ docs/source/api/discretisation/geometry/battery_geometry.rst
