+++ docs/source/api/discretisation/index.rst
@@ -1,0 +1,15 @@
+Discretisation
+==============
+
+The ``Discretisation`` class is used to discretise a model onto a mesh. Various spatial
+methods, geometries, and meshes are available.
+
+.. autoclass:: pybamm.Discretisation
+  :members:
+
+.. toctree::
+  :maxdepth: 2
+
+  geometry/index
+  meshes/index
+  spatial_methods/index
