+++ docs/source/api/discretisation/spatial_methods/index.rst
