+++ docs/source/api/discretisation/spatial_methods/scikit_finite_element.rst
