+++ docs/source/api/discretisation/spatial_methods/spatial_method.rst
