+++ docs/source/api/discretisation/spatial_methods/zero_dimensional_method.rst
