--- docs/source/api/spatial_methods/discretisation.rst
@@ -1,5 +1,0 @@
-Discretisation
-==============
-
-.. autoclass:: pybamm.Discretisation
-  :members:
