+++ pybamm/discretisation/meshes/one_dimensional_submeshes.py
