+++ pybamm/discretisation/spatial_methods/spatial_method.py
