+++ tests/unit/test_discretisations/test_geometry/test_battery_geometry.py
