+++ tests/unit/test_discretisations/test_spatial_methods/test_base_spatial_method.py
