+++ tests/unit/test_discretisations/test_spatial_methods/test_zero_dimensional_method.py
